+++ tests/testthat/test-dots.R
@@ -109,3 +109,30 @@ test_that("dots_node() doesn't trim attributes from arguments", {
   dots <- eval(expr(dots_node(!! x)))
   expect_identical(node_car(dots), x)
 })
+
+test_that("dots_definitions() uses tidy eval", {
+  var1 <- "foo"
+  var2 <- "bar"
+  dots <- dots_definitions(pat = foo(!!var1) := bar(!!var2))
+
+  pat <- list(lhs = quo(foo("foo")), rhs = quo(bar("bar")))
+  expect_identical(dots$defs$pat, pat)
+})
+
+test_that("dots_definitions() accepts other types of arguments", {
+  dots <- dots_definitions(A = a := b, B = c)
+  expect_identical(dots$defs$A, list(lhs = quo(a), rhs = quo(b)))
+  expect_identical(dots$dots$B, quo(c))
+})
+
+test_that("dots patterns check input types", {
+  expect_error(dots_patterns(foo()), "`...` can only contain two-sided formulas")
+  expect_error(dots_patterns(foo(), .operator = ":="), "`...` can only contain `:=` definitions")
+  expect_error(dots_patterns(foo(), .operator = "foo"), "`...` can only contain calls to `foo` with two arguments")
+})
+
+test_that("dots patterns accept arbitrary binary functions", {
+  pat <- list(A = list(lhs = quo(a), rhs = quo(b)))
+  expect_identical(dots_patterns(A = a + b, .operator = "+"), pat)
+  expect_identical(dots_patterns(A = foo(a, b), .operator = "foo"), pat)
+})